Rindsel-package {Rindsel}	R Documentation

R package for Selection Indices

Description

Estimation of marker effects for use in genomic selection and mate allocation.

Details

Package: 	 Rindsel

Type: 	 Package

Version: 	 1.3

Date: 	 09.05.2013

License: 	 GPL (version 2 or later)

LazyLoad: 	 yes

The following functions can be used for the estimation of Genomic Selection Index(GSI):

ESIMIndex: 	 Eiegen Selection Index Method (ESIM).

KNIndex: 	 Restrictive Kempthorne & Nordskog selection index.

LTIndex: 	 Lande and Thompson selection index.

MESIMIndex: 	 Molecular eigen selection index method.

RESIMIndex: 	 Restrictive eigen selection index method.

SmithIndex: 	 Smith Selection Index

KNGSIIndex: 	 Restrictive Kempthorne & Nordskog Genomic Selection Index.

RGESIMIndex: 	 Restrictive Genomic Eiegen Selection Index Method.

TGSIIndex: 	 Tallis predetermined proportional gains Genomic Selection Index.

TGESIMIndex: 	 Tallis predetermined proportional gains Eiegen Selection Index Method.

GESIMIndex: 	 Genomic Eiegen Selection Index Method.

SGSIIndex: 	 Smith Genomic Selection Index.

Statistical Metodology
Ceron-Rojas and Crossa (2011) made a detailed presentation of a multitrait GEBV method to obtain the statistical properties and its sample, also defined the genomic selection index (GSI). Here is a brief summary of this approach; next it is exposed the Mallard approximation (1972), and finally the coeficients of the predetermined proportional genomic gains and restricted selection indices are derived.
Breeding values within a multi-trait context
Consider a base population with genotypes and markers in the univariate context, the vector of breeding values (a) can be defined as a=Xu~MVN(0,XX'*sigma_{u}^{2}, where X=x_{ij} (i=1,...,g;j=1,...,m) is a matrix of marker covariables, x_{ij}, which could be 1, 0 or -1 for a bi-allelic genotypic codominant markers of an F_2 segregating population or 1 and -1 for a double haploid (DH) population of inbred lines; u~MVN(0,I_m*sigma_{u}^{2}is a vector mx1 of the genotypic marker effects, I_m is an identity matrix of order mxm, sigma_{u}^{2} is a constant variance for each marker effect, and XX'*sigma_{u}^{2}=A*sigma_{A}^{2}, , where A is the numerical relationship matrix of order gxg, and sigma_{A}^{2} is the additive genetic variance. The relationship between matrices AA' and A is E(XX')=alpha A+B (Habier et al., 2007), where alpha is a scalar and B is a matrix that depends on the marker allelic frequencies, and E(XX') is the expected value of matrix XX'; the difference alpha^{-1}XX'-alpha^{-1}B=Phi is called genomic relationship matrix.

Within a multi-trait context, the vector of breeding values (a) can be defined as a=Zu, where a=[a_{1}',a_{2}'...a_{t}'~MVN(0,U%x% I_{m})] is a (txg)x1 vector of t subvectors of breeding values of the g candidate for selection for t traits, Z=I_t %x% X is a block diagonal matrix (txg)x(txm) of marker covariables, G is the genetic covariance matrix; I_m, Phi and X were already defined for a single trait, I_t is an identity matrix of order txt; u'=[u_{1}',u_{2}'...u_{t}'~ MVN(0,U%x%I_{m})] is a (txm)x1 vector of t subvectors of the m genotypic marker effects for t traits, and U is a variance-covariance txt matrix of u; %x% denotes the Kronecker product.

Computing the GEBV within a multi-trait context
Suppose that the covariance for any pair of phenotypic means for traits q^{th} and q^{'th} (q, q'=1,2,...,t) is Cov(y_{q},y_{q'})=XX'*sigma_{u_{qq'}}+I_{g}*sigma_{e_{qq'}}, where sigma_{u_{qq'}} and sigma_{e_{qq'}} are constant covariances of each genotypic marker effect and the covariance of the residual effects for traits q^{th} and q^{'th}, respectively, the multi-trait model for predicting the breeding values of the g genotypes

candidates for selection is

y=Zu+epsilon

where y'=[y_{1}',y_{2}'...y_{t}']~MVN(0,G%x% Phi +R x I_{g}) is a is a (txg)x1 vector of t subvectors of phenotypic means of the g genotypes; Var(y)=G%x% Phi+R%x% I_{g}=V is the variance-covariances

matrix of y; G,Phi,u and Z were already defined; epsilon'=[e_{1}',e_{2}'...e_{t}']~ MVN(0,R%x%I_{g}) is a vector (txg)x1 of t subvectors of residuals for the t traits of interest, R is a variance-covariance matrix tx t of residual values of t traits, I_{g} is an identity matrix of order gxg. In this case epsilon is independent of u but the covariance between e_{q} and e_{q'}, and the covariance between u_{q} and u_{q'} (q,q'=1,2,...,t) are not zero, i.e., Cov(u_{q},u_{q'})=I_{m}*sigma_{qq'}.

hat{a}=E(a/y)=(G%x%Phi)V^{-}y

where hat{a}'=[hat{a}_{1}'hat{a}_{2}'...hat{a}_{t}'] is a (txg)x1 vector of t subvectors of GEBV of the g candidate for selection for t traits; E(a/y) is the expected conditional vector of breeding values, a'=[a_{1}'a_{2}'...a_{t}'], giving t subvectors of phenotypic means, y; Cov(a,y)=G%x%Phi is the variance covariances matrix of a and y; matrix V^{-} is a generalized inverse of V; I_{g},I_{t},V,G and Φ were already defined.

Genomic selection index (GSI)
To develop the proportional and restricted genomic selection indices, it is necessary first to redefine the PSI in the context of genomic selection. The PSI, I_{y}=∑^{t}_{q=1} B_{q}y_{q} is a linear combination of the mean phenotypic values of the traits of interest (y_{q}); in genomic selection, Ceron-Rojas and Crossa (2011) defined the genomic selection index as a linear combination of the GEBV of the t traits of interest using the vector hat{a}'=[hat{a}_{1}'hat{a}_{2}'...hat{a}_{t}'], i.e.,

I_{m}=∑^{t}_{q=1} B_{q}hat{a}_{q}

where B'=[B_{1}'B_{2}'...B_{t}'] denotes the proportion which the GEBV of the candidate for selection for t traits contributes to the genomic selection index. The difference between a PSI and a GSI is that y'=y_{1}y_{2}...y_{t} is replaced by hat{a}'=[hat{a}_{1}' hat{a}_{2}' ... hat{a}_{t}'].

It is possible to denote the net genetic merit or breeding value of the individual candidates for selection as θ'g, where g'=[g_{1} g_{2} ...g_{t}] is a vector of breeding values for t traits and θ'=[θ_{1} θ_{2} ...θ_{t}] is a vector of economic weight; the response to the GSI(R_{I_{M}}) is

R_{I_{M}}=kσ_{H}{θ'Γβ}/{√{θ'Gθ}√{β'Γβ}}=kσ_{H} ρ_{HI_{M}}

where k is the standardized selection differential, σ_{h}^{2}=θ G θ is the variance of σ_{I_{M}}^{2}=β'Γβ, is the variance of I_{M}, Cov(I_{M},H)=θ'Γβ, and ρ_{HI_{M}} are the covariance and the correlation between I_{M} and H, respectively; G is the genetic variance-covariance matrix, and Γ is the variance-covariance means matrix of the GEBV for t traits. In the following developments we extend the theory of Mallard (1972) and the approach proposed in Ceron-Rojas et al. (2008a)and Ceron-Rojas and Crossa (2011) for the proportional and restricted genomic case.

Restrictions on the Genomic Selection Indices
The proportional and restricted genomic selection indices exposition in the present work follows the Mallard (1972) and Itoh and Yamada (1987) ideas and it is shown that this approximation comprises the remaining SI except the Ceron-Rojas et al. (2008a) approximate, which is based on the canonical correlations.

Mallard Approximation to the Predetermined and Proportional Gains Selection Indices
Kempthorne and Nordskog (1959) maximized the square of the correlation between the genetic merit net and PSI under the constraint W'GB=0, where W is a matrix of 1's and 0's, B is the vector of coefficients of PSI and G is the matrix of genetic variances and covariances. Number 1 in the matrix indicates that the expected genetic advance in that trait is zero and 0 indicates that the trait is not restricted.

Tallis (1985) got the predetermined and proportional gains selection index by minimizing the variance of the difference between PSI and the genetic merit net under the constraint W'GB=τ c, where τ is the coefficient of proportionality, c'=[c_{1} c_{2} ... c_{r}] is the vector of proportional restrictions imposed by the breeder, and r is the number of constraints. The result of Tallis (1985) is equivalent to Harville (1975) who maximized the correlation between genetic merit net and PSI under the above restriction. Mallard (1972) showed that under the condition C'c=0, restrictions

C'W'GB=0

link RPSI with the predetermined and proportional gains PSI. Meanwhile, Itoh and Yamada (1987) have shown that the matrix C is not unique and that any matrix that satisfies the condition is permissible. One such matrix is

C'=

c_r 	 0 	 ... 	 0 	 -c_1

0 	 c_r 	 ... 	 0 	 -c_2

... 	 ... 	 ... 	 ... 	 ...

0 	 0 	 ... 	 c_r 	 -c_r-1

Matriz C' is (r-1)x r order and range (r-1), and c_{p}(p=1,2,...,r) is the p^{th} element of the vector of proportional constraints, c'=[c_1 c_2 ... c_r], imposed by the breeder.

Using Mallard approximation for the restricted and proportional genomic selection indices According to Ceron-Rojas and Crossa (2011), the expected genetic advance in the GSI is Δ bar{g}_{M}=k{Γβ}/{√{β'Γβ}}, where k is the selection intensity, Γ is the average matrix of variances and covariances of the GEBV and β is the vector of coefficients of GSI. Under the condition of Mallard (1972), C'c=0, one way to obtain the coefficient vector is minimizing the TGSI of the variance of the difference between the net genetic merit H=θ 'g of the candidate selection and GSI (I_{m}=∑^{t}_{q=1} β_{q} hat{a}_{q}), Var(I_{M}-H), ie, it is necessary to minimize

Ψ =Var(I_{M}-H)+2v'C'W'Γβ

respect to β and v, where v'=[v_1 v_2 ... v_{r-1}] is a vector of Lagrange multipliers. From where,
B=[I-WC(C'W'*Gamma W)^{-1}W'*Gamma]*theta

is the vector of coefficients of TGSI, which is similar to PSI of Tallis (1985). When C=I beta=[I-W(W'Gamma W)^{-1}W'Gamma]*theta

is the vector of coeficientes of KN GSI similar to RPSI of Kempthorne and Nordskog (1959), and when C=I and W is a null matrix, β=0, i.e., β is the coefficient vector of Smith GSI without any restriction(Ceron-Rojas and Crossa, 2011). Furthermore, if the correlation ρ_{HI_{M}} is maximized in the context of canonical correlation (Ceron-Rojas et al., 2008 a and b; Cer\'on-Rojas and Crossa, 2011) under the Mallard (1972) condition, C'c=0, i.e., deriving

Ψ=θ'Γβ-0.5λ(β'Γβ-1)-0.5μ(θ'Gθ-1)-v'C'W'Γβ

respect to β, θ, v, λ and μ are Lagrange multipliers, the result is

(Q-λ I)β=0

where Q=[I-WC(C'W'Γ WC)^{-1}C'W'Γ]G^{-1}Γ and λ=ρ^{2}_{HI_{M}}. In this case, β is the vector of coefficients TGESIM similar to TGSI, and when C=I, Q=[I-W(W'Γ W)^{-1}W'Γ]G^{-1}Γ, and in the case β is the vector of coefficients RGESIM similar to KN GSI; when C=I and W is a null matrix, Q=G^{-1}Γ, and in such case has the genomic eigen selection index method (GESIM) of Ceron-Rojas and Crossa (2011). Thus, for restricted and proportional change in GESIM (RGESIM and TGESIM), the value that maximizes ρ_{HI_{M}} is the square root of the first eigenvalue λ) of matrix Q, and the vector that allows constructing I_{M} (with maximum correlation to H) is the first eigenvector of matrix Q(β); the value of β depends on the context in which it is estimated.

The weights of RGESIM and TGESIM
Note that the vector of coefficients, β, RGESIM and TGESIM don't incorporate economic weights; this makes selection difficult as important economic traits may have smaller weights in the eigenvector than less important ones, for this reason, Ceron-Rojas and Crossa (2011) proposed a method to modify the proportion given by the eigenvector throughout a linear transformation of β. In accordance with Ceron-Rojas and Crossa (2011), an equivalent way to write Eq.(7) is

(DQD^{-1}-λ I)b=0 (8)

Where b=dβ is a linear transformation of β and D=d_{1} d_{2} ... d_{t} is a diagonal matrix of weights of β. In this case b incorporates into the RGESIM and TGESIM information on the breeding values through matrix G, and information on the GEBV variances and covariances through matrix Γ; also, matrix D allows modifying the proportions of B according to the researchers interest in the traits. Similar to Eqn (7), the value that maximizes ρ_{{HI}_{M}} in Eqn (8) is the first eigenvalue (λ) of matrix DQD^{-1}, which is not affected by matrix D (Wilkinson, 1965), and the vector that allows constructing (with maximum correlation to H) is the first eigenvalue (λ) of matrix DQD^{-1}, which is not affected by matrix D (Wilkinson, 1965), and the vector that allows constructing (with maximum correlation to H) is the first eigenvector of matrix DQD^{-1} (1), which is indeed affected by D. When D is an identity matrix, then Eqns (7) and (8) are the same. Ceron-Rojas and Crossa (2011) showed how to construct a matrix D such that when vector β is transformed, vector b will preserve the required proportions for each trait. Note that the equality dβ=b can be written as

DD_{β}=D_{b}

where D_{β}=β_{1} β_{2} ... β_{t} and D_{b}=b_{1} b_{2} ... b_{t} are diagonal matrices with elements of the vectors β and b; then D=D_{b}D_{β}^{-1}=β_{q}^{-1}b_q=D_{β}^{-1}D_{b}, q=1,2,...,t, is a diagonal matrix with β_{q}^{-1}b_q as the q^{th} element, i.e., d_q=β_{q}^{-1}b_q. The values of b are unknowns and must be given by the researcher. One way to provide values to the elements of b is by multiplying matrix D_{β} by a diagonal matrix, D_{c}=({c_{1} c_{2} ... c_{t})}, where, c_{q}, q=1,2,...,t, is the {q^{th}} value assigned. Thus let D_{b}D_{β^{-1}}=D_{c}, then D_{b}=D_{c}D_{β}, from where D=D_{β}^{-1}D_{c}D_{β} and thus matrix DQD^{-1} of Eqn (8) is equal to D_{β}^{-1}D_{c}D_{β}QD_{β}^{-1}D_{c}^{-1}D_{β}.

Author(s)

Sergio Perez-Elizalde and Jose Crossa Alejandro I. Aguirre-Salado Maintainer: Sergio Perez-Elizalde <sergiop@colpos.mx>

References

Ceron-Rojas, J.J., J. Sahagún-Castellanos, F. Castillo-Gonzalez, and A. Santacruz-Varela, 2006. A selection index method based on eigenanalysis. Crop Science 46: 1711-1721.
Ceron-Rojas, J.J., J. Sahagún-Castellanos, F. Castillo-Gonzalez, A. Santacruz-Varela and J. Crossa, 2008. A restricted selection index method based on eigenanalysis. Journal of Agricultural, Biological, and Environmental Statistics 13: 449-457.
Cunningham, M. D., Mace, J. W., and Peters, E. R. (1969).Clinical experience with phenobarbitone in icterus neonatorum. Lancet, 1, 550.
Bulmer, M. G. 1986. Sex ratio theory in geographically structured populations. Heredity, 56, 69-73.
Hazel, L.N. 1943 The genetic basis for constructing selection indexes. Genetics 28, 476.
Jansen RC. Studying complex biological systems using multifactorial perturbation. Nat Rev Genet 2003, 4: 145-151.
Kempthorne, O., and A. W. Nordskog, 1959. Restricted selection indices. Biometrics 15: 10-19.
Lande, R., and R. Thompson, 1990. Efficiency of marker-assisted in the improvement of quantitative traits. Genetics 124: 743-756.
Li Z 1998. Molecular analysis of epistasis affecting complex traits. In: Paterson AH (ed) Molecular dissection of complex traits. CRC Press, Boca Raton, pp 119-130
Muirhead, R. J., 2005. Aspects of Multivariate Statistical Theory, John Wiley and Sons, New York.
Smith, H. F., 1936. A discriminant function for plant selection, In: Papers on Quantitative Genetics and Related Topics. Deparment of Genetics, North Carolina State College,Raleigh, North Carolina, pp. 466-476.

[Package Rindsel version 1.3 Index]