BLPSI {Rindsel}	R Documentation

Base Linear Phenotypic Selection Index.

Description

This function estimates the BLPSI selection index.

Usage

BLPSI(datos,file.wgt,selval,design,corr,out="outextBLPSI.txt",outcsv="outBLPSI.csv",
rawdata,one.env,block.ex,softR,file.covG)

Arguments

datos
dataframe object which contains the traits (columns) and entries (rows)

file.wgt
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits.

selval
Percent of selected genotypes (5, 10, 20, etc.). Default is selval=5.

design
Field design ("lattice", "rcb", "AdjMeans"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="out.csv".

rawdata
If genotypic and phenotypic variances are already calculated, then it can be used to calculate directly the index by setting rawdata=FALSE. Default is TRUE.

one.env
Use FALSE for multienviromrent trials.

block.ex
Use FALSE always.

softR
Use "" always.

file.covG
When design is "AdjMeans" and rawdata is FALSE, write the location of your variance and covariance matrix csv file.

Details

To derive the LPSI theory, we assumed that the phenotypic (P) and the genotypic (G) covariance matrix, and the vector of economic values (w) are known. However, P,G and w are generally unknown and it is necessary to estimate them. There are many methods for estimating P and G (Lynch and Walsh 1998) and w (Cotterill and Jackson 1985; Magnussen 1990). However, when the estimator of P is not positive definite (all eigenvalues positive) or the estimator of G is not positive semidefinite (no negative eigenvalues), the estimator of G could be biased. In this case, the base linear phenotypic selection index (BLPSI):

I_B=w'y

may be a better predictor of H=w'g than the estimated LPSI I = b'y (Williams 1962a; Lin 1978) if indeed the vector of economic values is known. Many authors (Williams 1962b; Harris 1964; Hayes and Hill 1980, 1981) have investigated the influence of parameter estimation errors on LPSI accuracy and concluded that those errors affect the accuracy of I = b'y when the accuracy of P and G is low. If vector w values are known, the BLPSI has certain advantages because of its simplicity and its freedom from parameter estimation errors (Lin 1978). Williams (1962a) pointed out that the BLPSI is superior to I = \hatb'y unless a large amount of data is available for estimating P and G.

There are some problems associated with the BLPSI. For example, what is the BLPSI selection response and the BLPSI expected genetic gains per trait when no data are available for estimating P and G? The BLPSI is a better selection index than the standard LPSI only if the correlation between the BLPSI and the net genetic merit is higher than the correlation between the LPSI and the net genetic merit (Hazel 1943). But if estimations of P and G are not available, how can we obtain the correlation between the base index and the net genetic merit? Williams (1962b) pointed out that the correlation between the BLPSI and H=w'g can be written as

ρ_ HI_B = √ w'Gw/w'Pw

and indicated that the ratio ρ_HI_B/ρ_HI can be used to compare LPSI efficiency vs BLPSI efficiency; however, in the latter case we at least need to know the estimates of P and G, i.e., \hat{P} and \hat{G}.

In addition, Equation (1) is only an assumption, not a result, and implies that P and G are the same. That is, b=P^{-1}Gw=w only when P=G, which indicates that the BLPSI is a special case of the LPSI. Thus to obtain the selection response and the expected genetic gains per trait of the BLPSI, we need some information about P and G. Assuming that BLPSI is indeed a particular case of the LPSI, the BLPSI selection response and the BLPSI expected genetic gains per trait could be written as

R_B = k_I √ w'Pw

and

E_B = k_IGw / √ w'Pw

respectively. The parameters of Equations (3) and (4) were defined earlier.

There are additional implications if b=P^-1Gw. For example, if P=G, then ρ_HI_B = √ w'Gw / w'Pw and BLPSI heritability h^2_I_B = w'Gw / w'Pw are equal to 1. However, in practice, the estimated values of the ρ_HI_B are usually lower than the estimated values of the ρ_HI.

Value

The BLPSI selection index and an output file which contains the selected entries and other useful information.

Author(s)

Maintainer: Angela Pacheco <r.a.pacheco@cgiar.org> Sergio Perez-Elizalde <sergiop@colpos.mx> Gregorio Alvarado Jesus Ceron Francisco Rodriguez Jose Crossa <j.crossa@cgiar.org> Juan Burgueno <j.burgueno@cgiar.org>

References

Ceron-Rojas, J.J., J. Sahagún-Castellanos, F. Castillo-Gonzalez, and A. Santacruz-Varela, 2006. A selection index method based on eigenanalysis. Crop Science 46: 1711-1721.

Ceron-Rojas, J.J., J. Sahagún-Castellanos, F. Castillo-Gonzalez, A. Santacruz-Varela and J. Crossa, 2008. A restricted selection index method based on eigenanalysis. Journal of Agricultural, Biological, and Environmental Statistics 13: 449-457.

Cunningham, M. D., Mace, J. W., and Peters, E. R. (1969).Clinical experience with phenobarbitone in icterus neonatorum. Lancet, 1, 550.

Bulmer, M. G. 1986. Sex ratio theory in geographically structured populations. Heredity, 56, 69-73.

Hazel, L.N. 1943 The genetic basis for constructing selection indexes. Genetics 28, 476.

Jansen RC. Studying complex biological systems using multifactorial perturbation. Nat Rev Genet 2003, 4: 145-151.

Kempthorne, O., and A. W. Nordskog, 1959. Restricted selection indices. Biometrics 15: 10-19.

Lande, R., and R. Thompson, 1990. Efficiency of marker-assisted in the improvement of quantitative traits. Genetics 124: 743-756.

Li Z 1998. Molecular analysis of epistasis affecting complex traits. In: Paterson AH (ed) Molecular dissection of complex traits. CRC Press, Boca Raton, pp 119-130

Muirhead, R. J., 2005. Aspects of Multivariate Statistical Theory, John Wiley and Sons, New York.

Smith, H. F., 1936. A discriminant function for plant selection, In: Papers on Quantitative Genetics and Related Topics. Deparment of Genetics, North Carolina State College,Raleigh, North Carolina, pp. 466-476.

Examples

library(Rindsel)
datos<-data.frame(read.csv("https://raw.githubusercontent.com/RAngelaPG/RIndSel-R/main/data/C1_PSI_05_Phen.csv",header=T,na.strings=c(NA,".","-"))) #Raw data to analized.
datos$REP=as.factor(datos$REP)																		#transform to factor.
datos$Block=as.factor(datos$Block)																	#transform to factor.
datos$ENTRY=as.factor(datos$ENTRY)																	#transform to factor.
file.wgt<-"https://raw.githubusercontent.com/RAngelaPG/RIndSel-R/main/data/weights_C1_PSI.csv"     #name of the file where we write the economic weights and restrictions. 
selval<-5                                                                                    		#Selection intensity.
design<-"lattice"                                                                            		#Experimental design.
corr<-FALSE                                                                                  		#You can decide if you want to work with the correlation matrix instead of variance and covariance matrix.
rawdata<-TRUE                                                                                		#By default is TRUE when you are using design option "lattice" or "rcbd", use FALSE for design option "AdjMeans".
one.env<-TRUE                                                                                		#Use FALSE for multienviromrent trials.
block.ex<-FALSE                                                                              		#Use FALSE always.
softR<-""                                                                                    		#Use "" always.
file.covG<-""                                                                                		#When design is "AdjMeans" and rawdata is FALSE, write the location of your variance and covariance matrix csv file.
BLPSI(datos,file.wgt,selval,design,corr,out="outextBLPSI.txt",outcsv="outBLPSI.csv",rawdata,one.env,block.ex,softR,file.covG)
file.show("outextBLPSI.txt")

[Package Rindsel version 5.1 Index]