ESIMIndex {Rindsel}	R Documentation

Eiegen Selection Index Method (ESIM).

Description

This function estimates the selection index by the Eiegen Selection Index Method (ESIM).

Usage

ESIMIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows)

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the ESIM selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

The eigen selection index method (or MESIM for its acronym in English) is a selection method based on the IS Smith (1936) in which uses the theory of singular value decomposition (DS) to estimate the vector β that maximizes the correlation ρ_{ZY}. In this case the first eigenvector elements (β) of {P} ^{-1}{G} determine the proportion to each character contributes to IS Y ={β}'p , also the first eigenvalue of {P} ^{-1}{G } is used in the response to selection, such IS was proposed by Ceron-Rojas et al. (2008a, 2008b).

MESIM theory is directly related to the theory of the canonical correlations in which, according to Muirhead (2005), when a set of variables may be divided so natural two subsets of random variables, and what you want is maximize the correlation between the two subsets of variables, canonical correlation analysis is very efficient. Consider the set of phenotypic variables (p) and the set of genotypic variables (g) in the context the improvement of plant and animal, in this case must be, indeed, the vector of phenotypic variables, p , and the vector of genotypic variables, g , belong, naturally, two sets of variables. Analysis canonical correlations reduces the correlation between p and g to its simplest form by linear transformations of p and g , ie Y=β'p and Z={θ'g} .

Using a concept similar to that of Kempthorne and Nordskog (1959), Ceron-Rojas et al. (2008a, 2008b) maximized the response to selection (Equation 2) to maximize ρ_{YZ}^2 . Note that the variances Y=β'p and Z={θ'g} are constant each selection cycle, therefore, the selection of genotypes can be done using Y=β'p or Y/√{{β'}{Pβ}} . Because of this, by maximizing ρ_{YZ}^2 is possible to introduce the constraints β'P β= 1 and θ'G θ= 1 so that in MESIM is necessary to maximize

Φ = (θ'G β)^2 - μ (β'P β -1 ) - ω (θ'G θ -1 )

with respect to β , θ , μ , and ω , where β is the vector of coefficients Y=β'p , θ is the vector coefficients of Z =θ'g , and μ and ω are multipliers Lagrange. In MESIM the values of θ does not necessarily economic weights.

When Φ is derived with respect to β and θ And the result is equal to zero vector, we have

(θ 'G β) ^ 2 Gθ - μPβ = 0

(θ 'G β) ^ 2 G β - ωG θ =0

By the constraints β'P β=1 and θ'G θ=1 , when the last equation is multiplied by β' and the next equation is multiplied by θ' ,the result is ({θ}'G β)^2 = ω= μ . Therefore, μ maximizes ρ_{YZ}^2 under the constraints β'P β =1 and θ'G θ= 1

The next problem is to determine the vector β that allows building the IS Y =β'p that has maximum correlation with Z=θ'g . According to Ceron-Rojas et al. (2008a, 2008b), the β required is the solution to the following equation

( Q - μ I)β = 0,

where Q= P ^{-1} G . Thus, in MESIM , the value that maximizes ρ_{YZ}^2 under the constraints β'P β=1 and θ'G θ=1 is the first eigenvalue ( μ) matrix Q , and the vector allows to construct Y=β'p (with maximum correlation with Z=θ'g ) is the first eigenvector ( β ) Matrix Q

Value

The ESIM selection index and an output file which contains the selected entries and other useful information.

References

Ceron-Rojas, J.J., J. Sahagun-Castellanos, F. Castillo-Gonzalez, and A. Santacruz-Varela, 2006. A selection index method based on eigen analysis. Crop Science 46: 1711-1721.
Ceron-Rojas, J.J., J. Sahagun-Castellanos, F. Castillo-Gonzalez, A. Santacruz-Varela and J. Crossa, 2008. A restricted selection index method based on eigenanalysis. Journal of Agricultural, Biological, and Environmental Statistics 13: 449-457

Examples

#library(Rindsel)
#data(MARKERS_GSI)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#ESIMIndex(TRAIT_GSI,WEIGHTS_GSI)
#T

[Package Rindsel version 1.3 Index]