KNIndex {Rindsel}	R Documentation

Restrictive Kempthorne & Nordskog selection index

Description

Funtion for the restrictive Kempthorne & Nordskog selection index

Usage

KNIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) foat which contains the traits (columns) for entries (rows).

weights
Text file in comma separated values (*.csv) foat whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in foat *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the KN selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

Kempthorne and Nordskog (1959) proposed an index that incorporates restrictions based on a predeteined level and improvement developed the theory of restricted selection index (ISR), which maximizes the genetic progress of some average Phenotypic variables while the other half is maintained without exchange. According to Cunningham et al. (1970), it is considered that an IS is restricted if the average allows one or more variables phenotype remains unchanged (no genetic gain) while in other genetic gain is maximized average. In general, the Kempthorne and Nordskog ISR (1959) has similar properties to IS Smith (1936).

Consider the problem of increasing grain yield in corn while other phenotypic variables remain without exchange. In this situation Kempthorne and Nordskog (1959) (KN) proposed maximize ρ_{YZ}^2 incorporating constraints matrix genotypic variance-covariance. Bulmer (1980) showed that Kempthorne and Nordskog results (1959) also obtained maximize {θ}'G β . In the derivation of the ISR Kempthorne and Nordskog (1959) follows the approach of Bulmer (1980).

Suppose we have q variables and desired phenotypic improve the average of qr of them while r of them remain without average exchange, the index Y=β'p must maximize ρ_{YZ} ^ 2 under such conditions. The restrictions are introduced by through the matrix W of 0s and 1s, where 1s indicate that mean values of the variables remain unchanged, while the 0s indicate average changes in phenotypic variables, such so that W'G β =0 . Let C =GW , since ρ_{YZ} is invariant under changes of scale, can assume that β 'P β = 1 , so θ 'G β should be maximized under the constraint β'C = 0 and β' P β = 1 . Let τ and v'=≤ft [{v_1 ... v_r} \right] Lagrange multipliers. Should be maximized

Ψ =θ'G β - τ (β'P β-1) -v 'C' β

In deriving Ψ over β and equate the result to null vector is

θ'G-2 τ P β -v'C' =0

Thus, the vector of coefficients of KN that maximizes ρ_{YZ}^2 is β_{KN}=[I -P^{-1}C (C'P^{-1}C)^{-1}C']β_S , where P^{-1} is the inverse matrix P and β_S is the vector defined in Equation 3 (the subscript denotes Kempthorne and Nordskog KN). If A=[ I - P^{-1} C(C'P ^{-1}C)^{-1}C'] , then

β_{KN} =A β_S.

Value

The KN selection index and an output file which contains the selected entries and other useful infoation.

Examples

#library(Rindsel)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#KNIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI)
#T

[Package Rindsel version 1.3 Index]