LTIndex {Rindsel}	R Documentation

Lande and Thompson selection index

Description

Funtion for the Lande and Thompson selection index

Usage

LTIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows).

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the LT selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

Lande and Thompson (1990) established the theoretical basis of the molecular marker-assisted selection (Molecular Marker Assisted Selection or MAS) for quantitative trait selection using computer simulation studies. Considerations Additional theoretical contributed more fully to understanding of fundamental aspects regarding the development of MAS population type, size \~n or sample and number of markers molecular (MM) required in MAS (Zhang and Smith 1992).

The procedure of Lande and Thompson (1990) is based on a selection index which combines information with phenotypic information from the of loci quantitative characters (Quantitative Trait Loci , or QTL) associated with the MM. This is because it is not possible to identify all QTL affecting the nature of interest (Li, 1998), that is, unless all the QTLs affecting phenotypic variables of interest are identified, MAS phenotypic information should be combined with the effects of the QTL associated with MM, also known MQTL effects for ensure the efficiency of selection (Dekkers and Settar, 2004). In the construction of selection index Lande and Thompson (1990) requires: (1) identify the linkage between the GM and the QTL on a map of MMs, (2) estimate the effects MQTLs, and (3) combined with the effects MQTLs phenotypic information to classify individuals by a selection index, and subsequently develop lines, varieties or populations of interest. The effects MQTLs can be identified and estimated based on linkage disequilibrium created by crossing inbred lines or divergent populations (Jansen, 2003).

As an illustrative example, assume that in the maize genome has a linkage map with ten molecular markers and the interest lies in enhancing the improvement of plant height; Suppose also that five QTLs with additive effects α_1 , α_2 ,..., α_5 have been identified in the genome associated with plant height and that the five QTLs are distributed at random from different chromosomes, ie, a QTL was found by chromosome. In practice, the idea of Lande and Thompson (1990) is to build a record (or score) by multiplying the values additives of the QTLs for the coded values of molecular markers (MM) to they are bound, ie , s = x_1 α_1 + x_2 α_2 + ... + x_5 α_5 , Where x_i , i = 1,2, ..., 5, denotes the coded values of the MM, which depend on the type of population that is used for selection in a population F_2 the possible values x_i are 1, 0 and -1. In this situation, the IS is constructed as

Y_{LT}=β_p p + β_s s

where p denotes the height of plant, s=x_1 α_1 + x_2 α_2 + ... + x_5 α_5 and what is desired is to maximize the correlation between Y_{LT} and Z = θ_p g_p . In this case the vector can be shown that that maximizes the correlation between Y_{LT} y Z=θ _p g_p is beta_{LT} =T^{-1}Kθ_{LT} , where β '_{LT} =[β '& β'_s , θ'_{LT} =[-1 & 0 ],

T=

σ _p^2 	 σ _s^2

σ _s^2 	 σ _s^2

K=

σ _g^2 	 σ _s^2

σ _s^2 	 σ _s^2

where σ _p^2 is the phenotypic variance, σ_g^2 is the genotypic variance and σ_s^2 is the variance of s=x_1 α _1 +x_2 α _2 +...+x_5 α _5 .

The case is considered more of a character is straightforward and All that changes is the size of the matrix T and K.

Value

The LT selection index and an output file which contains the selected entries and other useful information.

Examples

#library(Rindsel)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#data(MARKERS_GSI)
#data(QTL_GSI)

#LTIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI,markers=MARKERS_GSI,qtl=QTL_GSI)
#T

[Package Rindsel version 1.3 Index]