MESIMIndex {Rindsel}	R Documentation

Molecular eigen selection index method

Description

Funtion for MESIM selection index

Usage

MESIMIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows).

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the MESIM selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

MESIM is a generalization of ESIM to the case in which is incorporated MM information to IS, similarly as in the IS Lande and Thompson (1990). Following the basic idea of Kempthorne and Nordskog (1959), Ceron-Rojas et al. (2008b), maximized the correlation between Y_M =β'_p*P +β'_s*S =β'_M*P_{ps} and Z_M =θ'_1*g +θ_2 *s=θ'_M*g_{gs} , ρ_{Y_M Z_M}^2 , where p and g have been defined in Equations 1 and 2, and s is the vector of records of the additive effects of QTLs associated with MM, P'_{ps}=[p' s'] , G'_{gs}=[g' s'] , β'_M=[β'_P β'_s] , and θ'_M=[θ'_1 θ'_2] . In the illustrative example of Lande and Thompson (1990), Y_M = β_p*p + β_s*s , and Z_M = θ_p*G_P + θ _s*s , where p denotes the plant height and s = x_1 α_1 + x_2 α_2 + ... + x_5 α_5.

Again, because ρ_{Y_M Z_M }^2 is invariant to changes in is maximized scale ρ_{Y_M Z_M ^2} under the constraints β '_M*T*β_M = 1 and θ '_M*K*θ_M = 1 , so in MESIM, it is necessary maximize

(θ '_M*K*β_M-1 )^2-μ(β '_M*T*β_M-1) -ω(θ '_M*K*θ_M-1 )

With respect to β_M , θ_M , μ , and ω , where β_M is the vector of coefficients MESIM , θ_M is the vector of coefficients Z_M =θ'_M* g_{gs} and μ and ω are Lagrange multipliers. Ceron-Rojas et al. (2008b) found that the solution is equal

(Q-μ I)β_M=0

where Q=T^{-1}*K . Thus, in MESIM , the value that maximizes ρ_{Y_M Z_M} ^ 2 under the constraints β '_M*T*β_M=1 and θ '_M*K*θ_M=1 is the first eigenvalue ( μ ) Matrix Q , and the vector for building Y_M (With maximum correlation with Z_M =θ'_M* g_{gs} is the first eigenvector ( β_M matrix Q .

Value

The MESIM selection index and an output file which contains the selected entries and other useful information.

Examples

#library(Rindsel)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#data(MARKERS_GSI)

#MESIMIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI,markers=MARKERS_GSI)
#T

[Package Rindsel version 1.3 Index]