RESIMIndex {Rindsel}	R Documentation

Restrictive eigen selection index method

Description

Funtion for restrictive eigen selection index method

Usage

RESIMIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows).

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the RESIM selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

Resim Selection Index (Eigen Restrictive Selection Index method)

RESIM is a generalization of ESIM to the case in which are incorporated restrictions on Y=β'p , the procedure for set the β is, therefore, very similar to ESIM and is based on IS restricted Kempthorne and Nordskog (1959). So, suppose you have q variables and phenotypic want to improve the means of q-r of them while r of them average remain unchanged, the index Y=β'p should maximize ρ_{YZ}^2 under conditions tale. Again, the restrictions are introduced by the matrix W of 0s and 1s, where 1s indicate that the mean values of the variables Phenotypic remain unchanged, while indicating 0s average changes in phenotypic variables, such that W'G β=0 . Let C= GW, since ρ_{YZ} is invariant under changes of scale can be assumed which β'*P*β=1 and θ'Gθ=1 , so ρ_{YZ}^2 be maximized under the constraints β'C=0 , β'*P*β=1 and θ'Gθ=1 . Let μ , ω and v'=[v_1 ... v_r] Lagrange multipliers. It should maximized

Φ = (θ '_M*K*β_M-1 )^2-μ(β '_M*T*β_M-1) -ω(θ '_M*K*θ_M-1 )-v'C'β

In deriving Φ over β , θ , μ , ω , and v'=[v_1 ... v_r] , Ceron-Rojas et al. (2008a) showed that

(Q-ω I)β=0

where now Q=[I-P^{-1}C(C'P^{-1}C)^{-1}P^{-1}G] . Thus, RESIM, the value that maximizes ρ _{YZ}^2 under the constraint β'P β = 1 , θ'G θ = 1 and β'C=0 is the first eigenvalue ( ω) of the matrix Q , and the vector lets build resim Y=β'p (with maximum correlation with Z=θ'g is the first eigenvector ( β matrix Q

Value

The RESIM selection index and an output file which contains the selected entries and other useful information.

Examples

#library(Rindsel)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#RESIMIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI)
#T

[Package Rindsel version 1.3 Index]