SmithIndex {Rindsel}	R Documentation

Smith Selection Index

Description

Funtion for Smith selection index

Usage

SmithIndex(file.dat = NULL, weights = NULL, selval = 5, design = "lattice", 
corr = FALSE, out = "out.txt", outcsv = "out.csv", rawdata = TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows).

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the Smith selection index for all the individuals.

rawdata
If genotypic and photypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

In 1936, Smith developed the methodology for the selection simultaneously on several variables based on the phenotypic linear combination of \

Y = β'p

Z = θ'g

Where Y is the selection index (SI), p = [p_{1} , p_{2} ,......, p_{q} ] and β = [b_{1} , b_{2} ,...... .., b_{q} ] is the vector of values Phenotypic and the coefficient vector of Y respectively, Z denotes the breeding value or net genetic component that can be gained through various selection cycles, g = [g_{1} , g_{2} ,......., g_{q} ] is the vector of values genotypic, and Θ = [ Θ_{1} , Θ_{2} ,..., Θ_{q} ] is the vector of economic weights, which according to Smith (1936), can be determined according to the experience of researchers and is regarded as a vector of constants. With the aim establish the relationship between phenotypic values P_{j} (j = 1, 2,......., q) observable and unobservable genotype in (Smith, 1936) proposed the following model

P_{j} = g_{j} + ε_{j}

Where g_{j} is the genotypic value of the jth feature and ε _{j} is the environmental component, which affects the genotypic character. (Smith, 1936) assumes that the interactions between g_{j } and \varepsilon _{j} can be considered as a random effect, therefore g_{j} has_{} only additive effects such as Z = Θ'g denotes the breeding value (Hazel, 1943; Kempthorne and Nordskog, 1959), under these assumptions, selection based on Y = β 'p leads to a response to selection (R) equal to

Where Σ and S are matrices variance-covariance matrix for phenotypic and genotypic values, respectively, k is the standard differential selection, θ ' Σ β is the covariance between Y, Z, θ 'Σθ is the variance of Z and β 'S β is the variance of Y, ρ _{yz} is the correlation between Y and Z. In the Smith selection index β _{s}=S^{-1}Σθ (where the subscript S denotes the method of Smith, S^{-1} is the inverse of the variance-covariance matrix of phenotypic S) allows us to construct the selection index (SI), Y =β_{s}'p, which maximizes the correlation with the breeding value.

Value

The Smith's selection index and an output file which contains the selected entries and other useful information.

Examples

#library(Rindsel)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#SmithIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI)
#T

[Package Rindsel version 1.3 Index]