TGESIMIndex {Rindsel}	R Documentation

Tallis predetermined proportional gains Eiegen Selection Index Method (TGESIM).

Description

This function estimates the selection index by the Tallis predetermined proportional gains Eiegen Selection Index Method (TGESIM).

Usage

TGESIMIndex (file.dat=NULL,weights=NULL,selval=5,design="lattice",corr=FALSE,
method=c("vanraden","classic"),out="out.txt",outcsv="out.csv",rawdata=TRUE)

Arguments

file.dat
Text file in comma separated values (*.csv) format which contains the traits (columns) for entries (rows)

weights
Text file in comma separated values (*.csv) format whose columns are: names of the traits, indicator variable for selected traits, economic weights or signs of the traits and indicator variable for restricted traits. If no name of the file is given, the file is choosen interactively.

selval
Percent of selected genotypes (5, 10, 20). Default is selval=5.

design
Field design ("lattice", "rcb"). Default is design="lattice".

corr
Selection index is based on variances-covariances matrix (FALSE) or correlation matrix (TRUE). Default is corr=FALSE.

method
Method for the computation of GEBV variance ("vanraden","classic")

out
Output file name. Default is out="out.txt".

outcsv
Output file in format *.csv. Default is outcsv="outcsv.csv". The file named "alloutcsv.csv" is generated, which contains the values of the traits and the ESIM selection index for all the individuals.

rawdata
If genotypic and phenotypic variances are already calculated, then it can be used to calculate directly the index by setting rawdat=TRUE. Default is FALSE.

Details

references at Rindsel-package

Value

The TGESIMIndex selection index and an output file which contains the selected entries and other useful information.

Examples

#data(MARKERS_GSI)
#data(TRAIT_GSI)
#data(WEIGHTS_GSI)
#TGESIMIndex(file.dat = TRAIT_GSI,weights=WEIGHTS_GSI,markers=MARKERS_GSI)
#T

[Package Rindsel version 1.3 Index]